requirements.txt
# Core ML
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
torch-geometric>=2.3.0

# AST Parsing
tree-sitter>=0.21.0
tree-sitter-python>=0.21.0

# Git & Data
gitpython>=3.1.0
tqdm>=4.65.0
networkx>=3.1

# App
streamlit>=1.28.0
pandas>=2.0.0
matplotlib>=3.7.0

# Utilities
numpy>=1.24.0
scikit-learn>=1.3.0
